Removed try catch statements and made code more reusable
